Move daemon files into daemon directory.
Rename Processor to DaemonDecider.
