+++ packages/cli/src/daemon/DaemonDecider.ts
@@ -22,15 +22,15 @@ import { DaemonClient } from "./DaemonClient";
 /**
  * Initial paramter parse to handle conditionally starting as a persistent process (e.g. daemon mode)
  * @export
- * @class Processor
+ * @class DaemonDecider
  */
-export class Processor {
+export class DaemonDecider {
 
     /**
      * Default port number
      * @private
      * @static
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private static readonly DEFAULT_PORT = 4000;
 
@@ -38,7 +38,7 @@ export class Processor {
      * Undocumented paramter for launching in server mode
      * @private
      * @static
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private static readonly DAEMON_KEY = "--daemon";
 
@@ -46,7 +46,7 @@ export class Processor {
      * Hold instance of a running server
      * @private
      * @type {net.Server}
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private mServer: net.Server;
 
@@ -54,7 +54,7 @@ export class Processor {
      * Hold current port number for the server
      * @private
      * @type {number}
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private mPort: number;
 
@@ -62,20 +62,20 @@ export class Processor {
      * Indicator for whether or not to start the server
      * @private
      * @type {boolean}
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private mStartServer: boolean;
 
     /**
-     * Creates an instance of Processor.
+     * Creates an instance of DaemonDecider.
      * @param {string[]} mParms
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     constructor(private mParms: string[]) { }
 
     /**
-     * Initialize our processor parse and optionally start the server
-     * @memberof Processor
+     * Initialize our DaemonDecider parse and optionally start the server
+     * @memberof DaemonDecider
      */
     public init() {
 
@@ -94,9 +94,9 @@ export class Processor {
     /**
      * Method to immediately parse or otherwise start the server for later processing from
      * incoming socket connections.
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
-    public process() {
+    public runOrUseDaemon() {
         if (this.mServer) {
             this.mServer.listen(this.mPort, "127.0.0.1", () => {
                 Imperative.api.appLogger.debug(`daemon server bound ${this.mPort}`);
@@ -110,7 +110,7 @@ export class Processor {
     /**
      * Server close handler
      * @private
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private close() {
         Imperative.api.appLogger.debug(`server closed`);
@@ -120,7 +120,7 @@ export class Processor {
      * Server error handler
      * @private
      * @param {Error} err
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private error(err: Error) {
         Imperative.api.appLogger.error(`daemon server error: ${err.message}`);
@@ -130,11 +130,11 @@ export class Processor {
     /**
      * Perform initial parsing of undocumented parameters
      * @private
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private initialParse() {
         const numOfParms = this.mParms.length - 2;
-        this.mPort = Processor.DEFAULT_PORT;
+        this.mPort = DaemonDecider.DEFAULT_PORT;
 
         if (numOfParms > 0) {
             const parm = this.mParms[2];
@@ -143,7 +143,7 @@ export class Processor {
              * NOTE(Kelosky): For now, we use an undocumented paramter `--daemon`.  If found first,
              * we bypass `yargs` and begin running this as a persistent Processor.
              */
-            const portOffset = parm.indexOf(Processor.DAEMON_KEY);
+            const portOffset = parm.indexOf(DaemonDecider.DAEMON_KEY);
 
             if (portOffset > -1) {
                 this.startServer = true;
@@ -162,7 +162,7 @@ export class Processor {
     /**
      * Get whether or not to start the server
      * @private
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private get startServer() {
         return this.mStartServer;
@@ -171,7 +171,7 @@ export class Processor {
     /**
      * Set whether or not to start the server
      * @private
-     * @memberof Processor
+     * @memberof DaemonDecider
      */
     private set startServer(startServer) {
         this.mStartServer = startServer;
